pyproject.toml
[project]
name = "krist-synth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "jinja2>=3.1.6",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "tqdm>=4.67.1",
]
